source_tree: add a default constructor for SourceTreeIterator
We simply use the default implementation.

In C++11, this constructor doesn't seem to have to create a valid iterator in
any way (it should just be assignable with operator=). In C++14, there are some
more conditions (it must behave like a past-the-end iterator on some unspecified
empty container and it must compare equal to other value-initialized iterators).